PORTOFOLIO
This portfolio project is a reflection of my journey, skills, and aspirations as a developer. Designed with simplicity and functionality in mind, the website provides a complete overview of who I am and what I bring to the table. It highlights my educational background, showcasing the foundation of my knowledge, and offers a detailed look into the projects I’ve worked on—each one a testament to my problem-solving abilities and passion for coding.

One of the key features of the portfolio is a dedicated section where I’ve outlined the programming languages and tools I’m proficient in, offering visitors a clear insight into my technical expertise. To enhance the user experience, I’ve incorporated a dark mode toggle, allowing users to switch between light and dark themes seamlessly. There’s also a convenient button to download my resume, making it easy for anyone to access a professional summary of my achievements.

In addition to these features, the portfolio serves as a gateway to connect with me on various platforms. With just a click, visitors can explore my work on GitHub, network with me on LinkedIn, or get a glimpse of my creative side on Instagram. This portfolio is more than just a website—it’s a blend of functionality and design, crafted to leave a lasting impression on anyone who visits.